# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Herstory AI - Backend REST API Specification",
    "description": "Herstory AI 백엔드 시스템 OpenAPI 3.0 명세서입니다.\n- Auth API: 로그인 & VIP 허브\n- Journey API: OCR 스캔, 기후 분석, Live Card 위젯, Apple Wallet .pkpass\n- Cart API: 스마트 장바구니 & ChoiceFit 피팅 신청\n- Store API: 면세점 체크인 & 매장 직원 태블릿 SSE 실시간 알림\n- Order API: 면세 결제 & Herstory Miles 적립\n- Care API: 현지 럭셔리 부티크 스팟, Google Maps, OpenAI 가죽 케어, 시티 스탬프, FCM\n- Flight API: 항공편 실시간 운항 정보 조회\n- Health API: 시스템 헬스 모니터링",
    "version": "v1.0.0",
    "contact": {
      "name": "Herstory Tech Team",
      "email": "tech@herstory.com"
    }
  },
  "servers": [
    {
      "url": "https://herstory-backend.onrender.com",
      "description": "Render Production Server (Live)"
    },
    {
      "url": "http://localhost:8080",
      "description": "Local Development Server"
    }
  ],
...
  "paths": {
...
    "/api/v1/store/check-in": {
      "post": {
        "tags": ["4. Store API (공항 면세점 체크인 & 태블릿 연동)"],
        "summary": "공항 면세점 체크인 처리",
        "description": "BLE/NFC 또는 수동 QR 체크인을 진행하고 매장 어시스턴트 태블릿 알림 및 웰컴 쿠폰을 전달합니다.",
        "operationId": "checkIn",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckInRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckInResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/care/visetos-spots": {
      "get": {
        "tags": ["6. Care API (현지 부티크 스팟 & 케어)"],
        "summary": "현지 글로벌 럭셔리 매장 스팟 탐색 및 가죽 케어 메시지 조회",
        "description": "목적지의 현지 럭셔리 플래그십 스토어 및 Care Desk 정보를 조회하고 맞춤형 케어 가이드를 반환합니다.",
        "operationId": "getVisetosSpots",
        "parameters": [
          {
            "name": "memberId",
            "in": "query",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CareResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/care/google-maps": {
      "get": {
        "tags": ["6. Care API (현지 부티크 스팟 & 케어)"],
        "summary": "SCR-502 Google Maps API 실시간 럭셔리 매장 탐색",
        "description": "Google Maps API를 활용해 현지 명품 부티크 위치 좌표 및 구글 지도 길안내 링크를 반환합니다.",
        "operationId": "getGoogleMapsSpots",
        "parameters": [
          {
            "name": "destination",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "default": "Bangkok" }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/VisetosSpot" }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/care/ai-care-tip": {
      "get": {
        "tags": ["6. Care API (현지 부티크 스팟 & 케어)"],
        "summary": "SCR-501 OpenAI 기반 가죽 케어 AI 가이드 생성 (다국어 지원: ko, en, ja, zh)",
        "description": "OpenAI GPT-4o를 활용해 구입 제품, 현지 기후 조건 및 다국어(ko, en, ja, zh) 맞춤형 가죽 관리 팁을 생성합니다.",
        "operationId": "getAiCareTip",
        "parameters": [
          {
            "name": "productName",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "default": "럭셔리 레더 백팩" }
          },
          {
            "name": "weather",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "default": "습도 88% 열대성 스콜" }
          },
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "default": "ko", "enum": ["ko", "en", "ja", "zh"] }
          }
        ],
        "responses": {
          "200": {
            "description": "성공 (텍스트 가이드)",
            "content": {
              "text/plain": {
                "schema": { "type": "string" }
              }
            }
          }
        }
      }
    },
    "/api/v1/care/push-test": {
      "post": {
        "tags": ["6. Care API (현지 부티크 스팟 & 케어)"],
        "summary": "FCM 디바이스 푸시 알림 테스트",
        "description": "Firebase Cloud Messaging(FCM) 푸시 알림 발송 테스트를 진행합니다.",
        "operationId": "testPushNotification",
        "parameters": [
          {
            "name": "title",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "body",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/care/stamp-checkin": {
      "post": {
        "tags": ["6. Care API (현지 부티크 스팟 & 케어)"],
        "summary": "SCR-502 현지 시티 패스포트 스탬프 획득 & 보상 적립",
        "description": "목적지 도시 럭셔리 부티크 방문 시 패스포트 스탬프를 획득하고 보너스 Herstory Miles를 적립합니다.",
        "operationId": "checkInCityStamp",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StampRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StampResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckInType": {
        "type": "string",
        "enum": ["BLE", "NFC", "QR", "MANUAL"]
      },
      "CheckInStatus": {
        "type": "string",
        "enum": ["COMPLETED", "PENDING", "EXPIRED"]
      },
      "PurchaseStatus": {
        "type": "string",
        "enum": ["PURCHASED", "PENDING_REENTRY", "ABANDONED"]
      },
...
      "CheckInRequest": {
        "type": "object",
        "required": ["memberId"],
        "properties": {
          "memberId": { "type": "integer", "format": "int64", "example": 1 },
          "checkInType": { "$ref": "#/components/schemas/CheckInType", "default": "BLE" },
          "qrCode": { "type": "string", "example": "HST-VIP-GATE24" }
        }
      },
      "CheckInResponse": {
        "type": "object",
        "properties": {
          "visitId": { "type": "integer", "format": "int64", "example": 101 },
          "memberId": { "type": "integer", "format": "int64", "example": 1 },
          "memberName": { "type": "string", "example": "김노마드 (VIP)" },
          "vipTier": { "type": "string", "example": "VIP" },
          "checkInType": { "$ref": "#/components/schemas/CheckInType" },
          "checkInStatus": { "$ref": "#/components/schemas/CheckInStatus" },
          "assistantNotified": { "type": "boolean", "example": true },
          "choiceFitRequested": { "type": "boolean", "example": true },
          "welcomeCouponMessage": { "type": "string", "example": "15% 즉시할인 웰컴 쿠폰 발급" },
          "purchaseStatus": { "$ref": "#/components/schemas/PurchaseStatus" },
          "visitedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "VisetosSpot": {
        "type": "object",
        "properties": {
          "spotName": { "type": "string", "example": "Herstory Siam Paragon Flagship" },
          "address": { "type": "string", "example": "991 Rama I Rd, Bangkok" },
          "locationType": { "type": "string", "example": "Herstory Flagship & VIP Lounge" },
          "latitude": { "type": "number", "example": 13.7466 },
          "longitude": { "type": "number", "example": 100.5349 },
          "careServiceAvailable": { "type": "string", "example": "가죽 방수 코팅 케어 제공" }
        }
      },
      "CareResponse": {
        "type": "object",
        "properties": {
          "destination": { "type": "string", "example": "Bangkok" },
          "pushNotificationMessage": { "type": "string", "example": "방콕 도착 환영 가이드" },
          "visetosSpots": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/VisetosSpot" }
          }
        }
      },
      "StampRequest": {
        "type": "object",
        "required": ["memberId", "spotName"],
        "properties": {
          "memberId": { "type": "integer", "format": "int64", "example": 1 },
          "spotName": { "type": "string", "example": "Herstory Siam Paragon" }
        }
      },
      "StampResponse": {
        "type": "object",
        "properties": {
          "memberId": { "type": "integer", "format": "int64", "example": 1 },
          "spotName": { "type": "string", "example": "Herstory Siam Paragon Flagship" },
          "cityName": { "type": "string", "example": "Bangkok" },
          "earnedMiles": { "type": "integer", "example": 1000 },
          "totalMiles": { "type": "integer", "format": "int64", "example": 69525 },
          "message": { "type": "string", "example": "+1,000 Herstory Miles 적립" }
        }
      },
      "PushResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "messageId": { "type": "string", "example": "projects/herstory-nomad-ai/messages/fcm-12345" },
          "title": { "type": "string", "example": "Herstory VIP 알림" },
          "body": { "type": "string", "example": "매장 방문을 환영합니다" },
          "statusMessage": { "type": "string", "example": "FCM 푸시 메시지 발송 완료" }
        }
      },
...
    }
  }
}